PR-B-6: extract LambdaBinder (function-literal binding + captured-variable analysis)
Move function-literal binding into a dedicated internal sealed
LambdaBinder, consolidating the captured-variable analysis and the
erased-adapter rewriter so the eventual #567 closure-emitter fix
(PR-E-9) has a single, clean BoundFunctionLiteralExpression contract
to assume.

Moved out of Binder.cs into the new src/Core/CodeAnalysis/Binding/LambdaBinder.cs:

  - BindFunctionLiteralExpression(FunctionLiteralExpressionSyntax)
  - CollectCapturedVariables(BoundStatement, ImmutableArray<ParameterSymbol>)
  - CreateErasedFunctionLiteralAdapter(BoundFunctionLiteralExpression, FunctionTypeSymbol)
  - GetErasedDelegateSlotType(TypeSymbol)        (private helper of the above)
  - TryGetFunctionLiteral(BoundExpression, out BoundFunctionLiteralExpression)
  - WrapAsTask(TypeSymbol)
  - nested class ErasedFunctionLiteralAdapterRewriter : BoundTreeRewriter
  - nested class CapturedVariableCollector       : BoundTreeRewriter

Binder constructs LambdaBinder once and routes the remaining call
sites (BindExpression dispatch, the BindCallExpression substitution
path, RebindFunctionLiteralDelegateArguments, and the BindTypeClause
async-return widening) through the new lambdas field.

TryGetFunctionLiteral remains accessible as
LambdaBinder.TryGetFunctionLiteral (public static) so Binder's
constructor keeps forwarding it as the OverloadResolver's
TryGetFunctionLiteralDelegate (the resolver consumes it from three
argument-shaping paths and does not gain a LambdaBinder reference).
CreateErasedFunctionLiteralAdapter and WrapAsTask are exposed as
public instance methods on LambdaBinder for the same reason.

Callbacks injected through the LambdaBinder constructor (no
back-reference to Binder):

  - bindBlockStatement, bindTypeClause, bindReturnTypeClause
  - isAsyncIteratorReturnType, resolveClrTypeForGenericArg
  - getCurrentFunction / setCurrentFunction (Binder.function is
    per-binder, not on BinderContext per the plan; BindFunctionLiteralExpression
    needs to push the synthetic function while binding the body)

ConversionClassifier.BindAndAttachParameterDefaultValue is the only
non-callback dependency and is consumed directly through the injected
ConversionClassifier instance.

RebindFunctionLiteralDelegateArguments stays on Binder for now — it
is called from non-lambda paths in the BindCallExpression /
BindUserInstanceCall families and will move with those in B-9
(ExpressionBinder); meanwhile it now uses LambdaBinder.TryGetFunctionLiteral
and lambdas.CreateErasedFunctionLiteralAdapter.

Validation:
  - Core.Tests (2120 passed, 1 skipped — RegenerateBaseline)
  - RefactoringBaselineTests: zero hash diffs across all curated samples
  - Compiler.Tests Sample_BuildsAndMatchesGolden (78 passed) —
    every sample, including the heavy function-literal ones
    (NamedDelegate, ClosureCaptureRefTypeField is intentionally
    skipped pre-PR for #567, others all green), produces byte-
    identical IL
  - Compiler.Tests Closure / Lambda / FunctionLiteral / Captured (41 passed)
  - Compiler.Tests Async / Func / Delegate emit subset (47 passed)
  - Compiler.Tests non-emit subset (121 passed)
  - Interpreter.Tests (72), LanguageServer.Tests (242), Sdk.Tests (24)

#567 is intentionally NOT fixed in this PR; the binder side is now
the single owner of function-literal binding, and the emitter-side
fix will land after E-9 ClosureEmitter is in place.

Next PR in the sequence is B-7 StatementBinder.

LoC: Binder.cs 14580 → 14242 (-338); LambdaBinder.cs new at 575.
